requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
PyJWT==1.7.1
requests==2.21.0
SQLAlchemy==1.3.1
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.15.0
